pyproject.toml
[tool.poetry]
name = "ck-be"
version = "0.1.0"
description = "A FastAPI backend for Coding Kittens"
authors = ["Matthew Aylward <matthewtyleraylward@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
redis-om = "^0.0.17"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.8"}
pydantic = {extras = ["email"], version = "^1.9.0"}
aioredis = "^2.0.1"
uvicorn = "^0.17.6"
pytest = "^7.1.1"
libcst = "^0.4.1"
requests = "^2.27.1"
pytest-cov = "^3.0.0"
coveralls = "^3.3.1"
jupyter = "^1.0.0"
dash = "^2.4.1"
dash-cytoscape = "^0.3.0"
jsonpickle = "^2.2.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
